Fixed IndexOutOfBounds error in Mesh.draw
First problem was that I had 12 extra faces in the new mesh
object; was caused by using a static faces list.

Main issue (and cause of the exception) was that I was trying
to access the 4th byte in the edgeDir array, but had only
initialised it to have 3 values.

'Cube' is now drawing to screen, but I still need to add the
code to Face.createNewFaces to determine the edge directions;
at the moment, I set them all to forward as default.